backend/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.10.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
motor==3.6.0
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.22
PyJWT==2.9.0
pymongo==4.9.0
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0

# WebSocket and Channels support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# HTTP client for AI service
aiohttp==3.10.11
